refactoring(run): refactoring of the return/throw position
