+++ src/main/java/fr/esgi/grp9/uparserbackend/run/service/IRunService.java
@@ -1,11 +1,14 @@
-package fr.esgi.grp9.uparserbackend.run.domain;
+package fr.esgi.grp9.uparserbackend.run.service;
+
+import fr.esgi.grp9.uparserbackend.run.domain.Run;
+import org.springframework.security.config.annotation.web.configurers.oauth2.server.resource.OAuth2ResourceServerConfigurer;
 
 import java.util.List;
+import java.util.Optional;
 
-public interface RunService {
+public interface IRunService {
     Run createRun(Run run);
-    Run findRunById(String id);
-    Run updateRun(Run run);
+    Optional<Run> findRunById(String id);
     List<Run> getRuns();
     void deleteRunById(String id);
 }
